The `connectTimeout` and `readTimeout` properties were specified as values of type `long` and while this worked for older versions of jersey, it no longer works with the version of jersey shipped with Payara 5. The values have been changed to be of type `int` that works across jersey versions.

+++ src/main/java/org/realityforge/keycloak/client/authfilter/Keycloak.java
@@ -166,13 +166,13 @@ public final class Keycloak
   private AccessTokenResponse callTokenService( @Nonnull final MultivaluedMap<String, String> parameters )
   {
     final ClientBuilder builder = ClientBuilder.newBuilder();
-    final long connectTimeoutInMillis = _config.getConnectTimeoutInMillis();
-    if ( connectTimeoutInMillis > 0L )
+    final int connectTimeoutInMillis = _config.getConnectTimeoutInMillis();
+    if ( connectTimeoutInMillis > 0 )
     {
       builder.property( "jersey.config.client.connectTimeout", connectTimeoutInMillis );
     }
-    final long readTimeoutInMillis = _config.getReadTimeoutInMillis();
-    if ( readTimeoutInMillis > 0L )
+    final int readTimeoutInMillis = _config.getReadTimeoutInMillis();
+    if ( readTimeoutInMillis > 0 )
     {
       builder.property( "jersey.config.client.readTimeout", readTimeoutInMillis );
     }
